chore: fix clang warnings (#2485)

+++ src/core/interpreter.cc
@@ -121,7 +121,7 @@ void RedisTranslator::OnDouble(double d) {
   ArrayPre();
 
   // Convert to integer when possible to allow converting to string without trailing zeros.
-  if (abs(fractpart) < kConvertEps && intpart < std::numeric_limits<lua_Integer>::max() &&
+  if (abs(fractpart) < kConvertEps && intpart < double(std::numeric_limits<lua_Integer>::max()) &&
       intpart > std::numeric_limits<lua_Integer>::min())
     lua_pushinteger(lua_, static_cast<lua_Integer>(d));
   else
